SlideChain/requirements.txt
web3
eth-account
python-dotenv
